+++ util/video.py
@@ -66,7 +66,7 @@ class SimpleDither(VideoDiff):
     def __mask(fframe, fprevframe, fill_value):
         # Mask frame over old frame
         # If element is different, change value to fill_value
-        masked_frame = np.where((fframe != fprevframe).any(axis=2, keepdims=True), [fill_value,fill_value,fill_value], fframe)
+        masked_frame = np.uint8(np.where((fframe != fprevframe).any(axis=2, keepdims=True), [fill_value,fill_value,fill_value], fframe))
         masked_frame = masked_frame.astype(np.uint8)
         return masked_frame
